DescargasOC-main/requirements.txt
PyPDF2
selenium
requests
requests-mock
python-dotenv
